+++ iam-login-service/src/main/java/it/infn/mw/iam/persistence/repository/client/IamClientRepository.java
@@ -32,8 +32,8 @@ public interface IamClientRepository extends PagingAndSortingRepository<ClientDe
 
   List<ClientDetailsEntity> findByClientNameLike(String clientName);
 
-  @Query("select c from ClientDetailsEntity c join ClientRelyingPartyEntity e where e.expiration < :dateTime")
-  List<ClientDetailsEntity> findByExpirationBefore(@Param("dateTime") Date dateTime);
+  @Query("select c from ClientDetailsEntity c join c.clientRelyingParty e where c.active = true and e.expiration < :dateTime")
+  List<ClientDetailsEntity> findActiveClientsExpiredBefore(@Param("dateTime") Date dateTime);
 
   @Query("select e.client from ClientRelyingPartyEntity e where e.entityId = :entityId")
   Optional<ClientDetailsEntity> findByEntityId(@Param("entityId") String entityId);